fix: circular dependencies and query invalidation
